Refactor include functions
